requirements.txt
Flask
PyPDF2
pdfplumber
pymongo
python-dotenv
pytesseract
pdf2image
Pillow
PyMuPDF